deal with review comment
